optimized codes of howTo for easier maintenance
