+++ src/screens/home/how-to/HowTo.tsx
@@ -1,96 +1,57 @@
-import { StepCardProps } from "./type";
-
-import { useEffect, useMemo, useRef, useState } from "react";
-
 import { PageContainer } from "@/components";
 import { Label } from "@/components/ui/label";
+import StepCardDesk from "./components/StepCardDesk";
 
-import StepCard from "./components/StepCard";
 import LocalizationKey from "@/i18n/key";
-import { useTranslation } from "react-i18next";
+import StepCardMobile from "./components/StepCardMobile";
+import { useStepReference } from "./utils";
 
 const HowTo = () => {
-  const [step, setStep] = useState(1);
-  const { t } = useTranslation();
-  const stepCard = useMemo<Omit<StepCardProps, "currentStep" | "onClick">[]>(
-    () => [
-      {
-        stepNumber: 1,
-        title: t(LocalizationKey.home.scan),
-        description: t(LocalizationKey.home.scanDescription),
-        imgClassName: "bg-[url('@/assets/timeline/timeline1.png')]",
-      },
-      {
-        stepNumber: 2,
-        title: t(LocalizationKey.home.topUp),
-        description: t(LocalizationKey.home.topUpDescription),
-        imgClassName: "bg-[url('@/assets/timeline/timeline2.png')]",
-      },
-      {
-        stepNumber: 3,
-        title: t(LocalizationKey.home.charge),
-        description: t(LocalizationKey.home.chargeDescription),
-        imgClassName: "bg-[url('@/assets/timeline/timeline3.png')]",
-      },
-      {
-        stepNumber: 4,
-        title: t(LocalizationKey.home.return),
-        description: t(LocalizationKey.home.returnDescription),
-        imgClassName: "bg-[url('@/assets/timeline/timeline4.png')]",
-      },
-    ],
-    [t],
-  );
-  const stepRefs = useRef<(HTMLDivElement | null)[]>([]);
-
-  useEffect(() => {
-    const observers = stepCard.map((_, index) => {
-      const observer = new IntersectionObserver(
-        (entries) => {
-          if (entries[0].isIntersecting) {
-            setStep(index + 1);
-          }
-        },
-        { threshold: 1 },
-      );
-
-      if (stepRefs.current[index]) {
-        observer.observe(stepRefs.current[index]!);
-      }
-
-      return observer;
-    });
-
-    return () => {
-      observers.forEach((observer) => observer.disconnect());
-    };
-  }, [stepCard]);
-
+  const { t, step, setStep, stepCard, stepRefsDesk, stepRefsMob } =
+    useStepReference();
   return (
     <PageContainer className="container mx-auto flex flex-col text-center">
-      <div className="space-y-14 lg:space-y-28">
+      <div className="space-y-14 lg:space-y-28 overflow-hidden">
         <Label
           variant={"banner"}
           className="bg-gradient-to-b from-green-900 to-green-600 bg-clip-text text-transparent"
         >
           {t(LocalizationKey.home.howtoUseIcharge)}
         </Label>
-        <div>
+        <div className="pointer-events-none hidden md:block md:translate-y-20 lg:translate-y-32">
           {stepCard.map((card, index) => (
-            <div
-              className="mx-auto flex w-[80%] grid-cols-9 flex-col md:grid md:w-[70%]"
-              ref={(el) => (stepRefs.current[index] = el)}
-              key={index}
-            >
-              <StepCard
-                stepNumber={card.stepNumber}
-                title={card.title}
-                description={card.description}
-                imgClassName={card.imgClassName}
-                currentStep={step}
-                onClick={() => setStep(card.stepNumber)}
-              />
-            </div>
+              <div
+                className="mx-auto grid w-[70%] grid-cols-5 flex-col"
+                ref={(el) => (stepRefsDesk.current[index] = el)}
+                key={index}
+              >
+                <StepCardDesk
+                  stepNumber={card.stepNumber}
+                  title={card.title}
+                  description={card.description}
+                  imgClassName={card.imgClassName}
+                  currentStep={step}
+                  onClick={() => setStep(card.stepNumber)}
+                />
+              </div>
+          ))}
+        </div>
+        <div className="pointer-events-none block translate-y-20 md:hidden">
+          {stepCard.map((card, index) => (
+              <div
+                className="ml-10 flex w-[60%] flex-col"
+                ref={(el) => (stepRefsMob.current[index] = el)}
+                key={index}
+              >
+                <StepCardMobile
+                  stepNumber={card.stepNumber}
+                  title={card.title}
+                  description={card.description}
+                  imgClassName={card.imgClassName}
+                  currentStep={step}
+                  onClick={() => setStep(card.stepNumber)}
+                />
+              </div>
           ))}
         </div>
       </div>
