+++ src/screens/home/how-to/components/HowToCard.tsx
@@ -3,25 +3,21 @@ import { HowToCardProps } from "../type";
 
 const HowToCard = ({
   className,
-  even = false,
   imgClassName,
-  children,
 }: Omit<HowToCardProps, "onClick">) => {
   return (
     <div
       className={cn(
-        "rounded-eyedrop aspect-square w-full p-5 shadow-slate-200",
+        "aspect-square w-full rotate-45 rounded-full p-5 shadow-slate-200 md:p-5 lg:p-10",
         className,
       )}
     >
       <div
         className={cn(
-          `aspect-square w-full rotate-[135deg] rounded-full border-2 text-black ${even ? "md:rotate-[-225deg]" : "md:rotate-[-45deg]"} shadow-inner-large bg-cover bg-center bg-no-repeat p-0`,
+          `shadow-inner-large aspect-square w-full -rotate-45 rounded-full border-2 bg-cover bg-center bg-no-repeat p-0 text-black`,
           imgClassName,
         )}
-      >
-        {children}
-      </div>
+      />
     </div>
   );
 };
